backend/requirements.txt
# Flask 核心框架
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0


# 数据库相关
PyMySQL==1.1.0

# 工具库
python-dotenv==1.0.0

# HTTP 客户端
requests==2.31.0

# MCP 协议支持 (官方SDK)
mcp

# OAuth 和 JWT 支持
Flask-JWT-Extended==4.6.0
cryptography>=42.0.0
PyJWT==2.8.0
authlib==1.3.0

# 生产环境必需依赖
gunicorn==21.2.0
gevent==23.9.1

# 后续可添加的依赖 (需要时取消注释)
# marshmallow==3.20.2
# marshmallow-sqlalchemy==0.29.0
# python-dateutil==2.8.2
# Flask-SocketIO==5.3.6
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0
# bcrypt==4.1.2
# Pillow>=10.0.0
# pydantic==2.5.0
# structlog==23.2.0
# Flask-Caching==2.1.0
# flasgger==0.9.7.1
# Flask-Mail==0.9.1
